Improve stability of CI jobs
The CI checks sometimes fail unexpectedly because of the 3rd party
dependeny check. This happens when a newer version of a crate we
depend on gets released which has not been vetted by ClearlyDefined
and/or the Eclipse Foundation.

In order to prevent this from happening, the cargo.lock file has been
added to version control in order to always use the same versions of
crates when running the checks.

Additionally, a job has been added to the nightly CI workflow to
verify that the library can be built with the latest versions.

Also removed unused tokio features to reduce dependeny footprint.