Keep track of reestablishment of subscriptions
Added a flag to the ConnectionState that indicates if all existing
subscriptions have been reestablished after connecting to the broker.

The reconnect function has been adapted to set this field based on
the outcome of its attempt to reestablish the subscriptions.

The is_connected function has been adapted to also consider this
newly added flag.